Update README with a brief professional introduction
Reduced the clutter in the README and made a clear, concise introduction about my professional persona. This commit was made to give visitors an immediate understanding about who I am as a software developer, and which technologies I am knowledgeable in. Additionally, the information is now neatly organized in segments, each having their own headlines for easier navigation and accessibility.